move GetVar* helper methods to Values object;
